api/requirements.txt
appdirs
black
click
dnspython
email-validator
fastapi
flake8
h11
idna
mccabe
motor
mypy-extensions
pathspec
pycodestyle
pydantic
pyflakes
pymongo
regex
starlette
toml
typed-ast
typing-extensions
uvicorn
gunicorn
pycountry
httpx
openpyxl
lxml
cssselect
python-multipart
python-jose[cryptography]
pytest
requests
fastapi-cache2
jinja2
